requirements.txt
streamlit==1.52.1

tensorflow==2.19.0
numpy==1.26.4

opencv-python-headless==4.8.1.78
Pillow==12.0.0

matplotlib==3.10.8
scikit-learn==1.7.2

seaborn